resize: Move handler to new resize_handler module.
